Refactor and add docs
