+++ reqwest-tracing/src/root_span_builder.rs
@@ -1,0 +1,101 @@
+use reqwest::{
+    header::{HeaderMap, HeaderValue},
+    Request, Response, StatusCode as RequestStatusCode,
+};
+use reqwest_middleware::{Error, Result};
+use task_local_extensions::Extensions;
+use tracing::Span;
+
+use crate::root_span;
+
+/// `RootSpanBuilder` allows you to customise the span attached by
+/// [`TracingMiddleware`] to incoming requests.
+///
+/// Check out [`root_span`] documentation for examples.
+///
+/// [`TracingMiddleware`]: crate::middleware::TracingMiddleware
+pub trait RootSpanBuilder {
+    fn on_request_start(req: &Request, extension: &mut Extensions) -> Span;
+    fn on_request_end(span: &Span, outcome: &Result<Response>, extension: &mut Extensions);
+}
+
+/// The default [`RootSpanBuilder`] for [`TracingMiddleware`].
+///
+/// [`TracingMiddleware`]: crate::middleware::TracingMiddleware
+pub struct DefaultRootSpanBuilder;
+
+impl RootSpanBuilder for DefaultRootSpanBuilder {
+    fn on_request_start(req: &Request, _extension: &mut Extensions) -> Span {
+        root_span!(req)
+    }
+    fn on_request_end(span: &Span, outcome: &Result<Response>, _extension: &mut Extensions) {
+        match outcome {
+            Ok(response) => {
+                // The request ran successfully
+                let span_status = get_span_status(response.status());
+                let status_code = response.status().as_u16() as i64;
+                let user_agent = get_header_value("user_agent", response.headers());
+                if let Some(span_status) = span_status {
+                    span.record("otel.status_code", &span_status);
+                }
+                span.record("http.status_code", &status_code);
+                span.record("http.user_agent", &user_agent.as_str());
+            }
+            Err(e) => {
+                // The request didn't run successfully
+                let error_message = e.to_string();
+                let error_cause_chain = format!("{:?}", e);
+                span.record("otel.status_code", &"ERROR");
+                span.record("error.message", &error_message.as_str());
+                span.record("error.cause_chain", &error_cause_chain.as_str());
+                if let Error::Reqwest(e) = e {
+                    span.record(
+                        "http.status_code",
+                        &e.status()
+                            .map(|s| s.to_string())
+                            .unwrap_or_else(|| "".to_string())
+                            .as_str(),
+                    );
+                }
+            }
+        }
+    }
+}
+
+fn get_header_value(key: &str, headers: &HeaderMap) -> String {
+    let header_default = &HeaderValue::from_static("");
+    format!("{:?}", headers.get(key).unwrap_or(header_default)).replace('"', "")
+}
+
+/// HTTP Mapping <https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/trace/semantic_conventions/http.md#status>
+///
+/// Maps the the http status to an Opentelemetry span status following the the specified convention above.
+fn get_span_status(request_status: RequestStatusCode) -> Option<&'static str> {
+    match request_status.as_u16() {
+        // Span Status MUST be left unset if HTTP status code was in the 1xx, 2xx or 3xx ranges, unless there was
+        // another error (e.g., network error receiving the response body; or 3xx codes with max redirects exceeded),
+        // in which case status MUST be set to Error.
+        100..=399 => None,
+        // For HTTP status codes in the 4xx range span status MUST be left unset in case of SpanKind.SERVER and MUST be
+        // set to Error in case of SpanKind.CLIENT.
+        400..=499 => Some("ERROR"),
+        // For HTTP status codes in the 5xx range, as well as any other code the client failed to interpret, span
+        // status MUST be set to Error.
+        _ => Some("ERROR"),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn get_header_value_for_span_attribute() {
+        let expect = "IMPORTANT_HEADER";
+        let mut header_map = HeaderMap::new();
+        header_map.insert("test", expect.parse().unwrap());
+
+        let value = get_header_value("test", &header_map);
+        assert_eq!(value, expect);
+    }
+}
